pyproject.toml
[project]
name = "blog-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-community>=0.3.29",
    "langchain>=0.3.27",
    "langgraph>=0.6.7",
    "python-dotenv>=1.1.1",
    "langchain-core>=0.3.76",
    "langchain-groq>=0.3.8",
    "fastapi>=0.116.1",
    "uvicorn>=0.35.0",
    "pydantic>=2.11.9",
    "watchdog>=6.0.0",
    "langgraph-cli[inmem]>=0.4.2",
]
